pyproject.toml
[tool.poetry]
name = "{{ project_name }}"
version = "0.1.0"
description = ""
authors = ["Nursultan Kudaibergenov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.5"
django = "3.2"
djangorestframework = "3.11.0"
fcm-django = "^1.0.12"
gunicorn = "20.0.4"
firebase-admin = "5.4.0"
psycopg2-binary = "2.9.5"
django-3-jet = "^1.0.7"
drf-spectacular = "^0.24.2"
django-silk = "^5.0.2"
django-filter = "^22.1"
django-environ = "^0.9.0"
django-solo = "^2.0.0"
pillow = "^9.3.0"
python-barcode = "^0.14.0"
huey = "^2.4.4"
redis = "^4.4.0"
pytelegrambotapi = "^4.8.0"
sentry-sdk = "^1.15.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pytest = "^7.2.0"
pytest-env = "^0.8.1"
pre-commit = "^2.20.0"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"
pytest-xdist = "^3.1.0"
deepdiff = "^6.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'core.settings'
python_files = 'test_*.py'
env = [
    "APP_DEBUG=True",
    "HUEY_IMMEDIATE=True"
]
addopts = "-p no:warnings"

[tool.black]
line-length = 100
py38 = true
include = '\.py$'

[tool.isort]
line_length = 100
force_single_line = "True"
atomic = "True"
lines_after_imports = 2
force_sort_within_sections = "True"
